requirements.txt
#Python 3.11.9
fastapi
sqlalchemy
psycopg2-binary
python-multipart
uvicorn
Authlib
python-jose
passlib
httpx
python-dotenv
itsdangerous
bcrypt